Implementing Inversion of Control
We create a proxy Interface (IDriver) between the application and a "provider". This provider is given control over the creation of the driver effectively inverting control.

The application now no longer has a dependency on the Third Party drivers however, the middle man still does and--although this can be otherwise avoided--the application needs to be re-compiled for every new customer.

Via Dependency Injection (a form of Inversion Control), we can make it such that, neither de application nor the Inverter have a hard dependency on them.