Dependency Injection
A class was created which reads the application settings and given that information, injects the "right" dependencies to the application. Note: In .NET, the application settings are stored in the app.config file as such, one is able to modify these settings without compiling.

Now, the application only depends on the Inverter library; the Inverter library has no dependencies; the proxy classes depend on the Inverter and the Third Party Drivers.

With this approach, if a new Data Store driver needs to be supported only a new proxy class gets created and the application settings are modified at the deployment system.